refactor: [#4] preview를 mobile사이즈로 제한하기 위해 decorators 적용
